refactor(mjh/resume): split session_service.py into lifecycle / turn / helpers modules
session_service.py (795 LOC) extracted into three focused modules:
- session_lifecycle_service: start_session, get_session_state, complete_session, _build_renderer_input
- session_turn_service: accept_pending, accept_custom, request_alternative, skip_target, navigate
- session_helpers: _with_turns, _load_active, _build_prior_context, _current_target,
  _apply_rewrite, _generate_next_proposal, _prefetch_all_proposals

session_service.py becomes a thin re-export shim; all existing import paths preserved.
Tests updated to import helpers from session_helpers and patch via the correct module namespace.
