+++ apps/myjobhunter/backend/tests/test_resume_refinement_proposal_cache.py
@@ -30,6 +30,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 import pytest
 
 from app.services.resume_refinement import session_service
+from app.services.resume_refinement import session_helpers
 
 
 def _fake_session(
@@ -131,7 +132,7 @@ async def test_navigate_cache_hit_skips_claude_call() -> None:
         return session
 
     with (
-        patch.object(session_service, "_load_active", new=fake_load_active),
+        patch.object(session_helpers, "_load_active", new=fake_load_active),
         patch.object(
             session_service.session_repo,
             "set_target_index",
@@ -182,7 +183,7 @@ async def test_navigate_cache_miss_falls_through_to_generation() -> None:
     generate_mock = AsyncMock(return_value=session)
 
     with (
-        patch.object(session_service, "_load_active", new=fake_load_active),
+        patch.object(session_helpers, "_load_active", new=fake_load_active),
         patch.object(
             session_service.session_repo,
             "set_target_index",
@@ -193,7 +194,7 @@ async def test_navigate_cache_miss_falls_through_to_generation() -> None:
             "hydrate_pending_from_cache",
             new=AsyncMock(return_value=None),  # cache miss
         ),
-        patch.object(session_service, "_generate_next_proposal", new=generate_mock),
+        patch.object(session_helpers, "_generate_next_proposal", new=generate_mock),
         patch.object(
             session_service.session_repo,
             "get_with_turns_for_user",
